near_swarm/core/market_data.py
"""
Market data integration for NEAR Protocol.
"""

import asyncio
from typing import Dict, List, Optional
import aiohttp
import time


class MarketDataManager:
    """Manages market data integration."""

    def __init__(self):
        """Initialize market data manager."""
        self.session = None
        self.price_cache = {}
        self.cache_expiry = 60  # 60 seconds

    async def initialize(self):
        """Initialize async resources."""
        if self.session is None:
            self.session = aiohttp.ClientSession()

    async def get_token_price(
        self,
        token: str,
        quote_currency: str = "USD"
    ) -> float:
        """Get token price from API."""
        await self.initialize()

        # Check cache
        cache_key = f"{token.lower()}_{quote_currency.lower()}"
        if cache_key in self.price_cache:
            timestamp, price = self.price_cache[cache_key]
            if time.time() - timestamp < self.cache_expiry:
                return price

        try:
            async with await self.session.get(
                f"https://api.coingecko.com/api/v3/simple/price",
                params={
                    "ids": token.lower(),
                    "vs_currencies": quote_currency.lower()
                }
            ) as response:
                data = await response.json()
                price = data[token.lower()][quote_currency.lower()]

                # Update cache
                self.price_cache[cache_key] = (time.time(), price)
                return price

        except Exception as e:
            raise Exception(f"Error getting {token} price: {str(e)}")

    async def analyze_market_opportunity(
        self,
        tokens: List[str],
        amount: float
    ) -> Dict:
        """Analyze market opportunities."""
        # Validate amount
        if amount <= 0:
            raise ValueError("Amount must be positive")

        try:
            # Get market state
            market_state = await self._get_market_state(tokens)

            # Find opportunities
            opportunities = await self._find_opportunities(
                market_state,
                amount
            )

            return {
                "timestamp": time.time(),
                "market_state": market_state,
                "opportunities": opportunities
            }

        except Exception as e:
            raise Exception(f"Error analyzing market: {str(e)}")

    async def _get_market_state(self, tokens: List[str]) -> Dict:
        """Get current market state."""
        try:
            # Get prices
            prices = {}
            for token in tokens:
                prices[token] = await self.get_token_price(token)

            # Get volumes
            volumes = await self._get_trading_volumes(tokens)

            # Get liquidity
            liquidity = await self._get_liquidity_data(tokens)

            return {
                "timestamp": time.time(),
                "prices": prices,
                "volumes": volumes,
                "liquidity": liquidity
            }

        except Exception as e:
            raise Exception(f"Error getting market state: {str(e)}")

    async def _get_trading_volumes(self, tokens: List[str]) -> Dict:
        """Get trading volumes."""
        # Mock implementation
        return {
            token: 1000000.0 for token in tokens
        }

    async def _get_liquidity_data(self, tokens: List[str]) -> Dict:
        """Get liquidity data."""
        # Mock implementation
        return {
            token: {
                "total_liquidity": 1000000.0,
                "available_liquidity": 500000.0
            }
            for token in tokens
        }

    async def _find_opportunities(
        self,
        market_state: Dict,
        amount: float
    ) -> List[Dict]:
        """Find trading opportunities."""
        opportunities = []

        for token, price in market_state["prices"].items():
            liquidity = market_state["liquidity"][token]
            profit = self._estimate_profit(price, amount, liquidity)

            if profit > 0:
                opportunities.append({
                    "token": token,
                    "type": "buy",
                    "price": price,
                    "estimated_profit": profit
                })

        return opportunities

    def _estimate_profit(
        self,
        price: float,
        amount: float,
        liquidity: Dict
    ) -> float:
        """Estimate potential profit."""
        # Mock implementation
        available_ratio = liquidity["available_liquidity"] / liquidity["total_liquidity"]
        return amount * price * 0.01 * available_ratio

    async def close(self):
        """Clean up resources."""
        if self.session:
            await self.session.close()
            self.session = None
